'FinFixed'

+++ app/main.py
@@ -7,22 +7,17 @@ def move_file(command: str) -> None:
 
     cmd, source_file, final_file = command.split()
 
-    if cmd == "mv":
+    if cmd == "mv" and source_file != final_file:
 
         with open(source_file) as source:
             data = source.read()
 
         final_path, fin_file_name = os.path.split(final_file)
 
-        if len(final_path) > 1:
-            directory = final_path
-        else:
-            directory = "."
+        if final_path:
+            os.makedirs(final_path, exist_ok=True)
 
-        if not os.path.exists(directory):
-            os.makedirs(directory, exist_ok=True)
-
-        with open(os.path.join(directory, fin_file_name), "w") as target_file:
+        with open(final_file, "w") as target_file:
             target_file.write(data)
 
         os.remove(source_file)
